+++ core/js/typespecs/prompt.ts
@@ -26,6 +26,8 @@ export {
   type MessageRole,
 } from "./openai/messages";
 
+export { anthropicMessageParamSchema } from "./anthropic/messages";
+
 export { toolsSchema } from "./openai/tools";
 export type { Tools } from "./openai/tools";
